Changed formatting of metadata
Fixed not properly writing metadata to archive and added/changed how ComicInfo.xml is formatted. Uses SeriesGroup instead of Series now as I think Magazines/Events are less a discrete series and more a grouping of series. This also improves how Komga sorts series vs. series groups. Added support for scraping months/years if applicable and now writes the URL used to scrape to the Web tag and the Parody to the AlternateSeries tag.